updated loaihang
